Make the lint gate real
`select = ["ALL"]` sat under `[tool.ruff]`, where modern ruff ignores it and
prints a deprecation warning on every run, so the gate this project declared was
never actually on. Turning it on as written would not have helped either: `ALL`
enables rules that contradict each other and rules that fight pytest, and stood
at 923 violations. The set here is the useful core -- correctness, imports,
naming, pyupgrade, bugbear, simplification, annotations, docstrings -- chosen for
the one property a gate needs, which is that it can be held at zero.

923 -> 0. Most of it was mechanical, but three findings were not:

`from __future__ import annotations` sat ABOVE the module docstring in protocol.py
and web_runner.py. That makes the string an ordinary expression rather than a
docstring, so both modules had no __doc__ at all and every import after it
counted as "not at top of file" -- all 22 E402s were this one mistake, which the
misconfigured linter had been hiding.

`l` was the waterline length throughout basic_hull_model, which reads as a 1. It
is now `length`, matching the signature the registered friction laws already use.
Two rotation matrices and an apparent-wind speed in hybrid_sail were similarly
renamed to match the lowercase convention basic_sail already follows.

hull_envelope.section unpacked `col_max` and never used it. That one needed
checking rather than deleting: the submerged area is integrated from the bottom
surface and the beam from the occupied-cell count, so the deck really is
irrelevant below the waterline.

Where a rule is wrong for this codebase it is ignored with the reason, not
worked around. RUF001 calls m², kg·m² and "9 × 12" ambiguous; they are meant.
Test docstrings are noun phrases saying what a fixture is, so D401 would reword
every one of them into something worse. Two deferred class imports carry a noqa
because they are classes, and PascalCase is correct.

FIXLIST.md comes across from chore/config-hygiene, where it was stranded on a
branch nothing descended from, with a dated banner noting what has landed. Its
original analysis is untouched. The CI workflow it calls for follows in its own
commit, which needs a token carrying the `workflow` scope to push.

381 tests pass and the physics is unchanged: flingo still builds ORCWithJibSail
on a Hughes hull at Cf 0.00390, added mass 1.35 / 26.46 / 1.83.

+++ test/test_basic_sail.py
@@ -19,20 +19,28 @@ def make_state(u: float = 0.0, v: float = 0.0, psi: float = 0.0) -> State:
 
 def make_sail(**overrides: float) -> BasicSail:
     """Build a sail with a fixed wind."""
-    return BasicSail({
-        "airfoil_name": "NACA0012", "area": 1.971, "alpha_min": -179, "alpha_max": 179,
-        "res": [1.8e5], "wind_speed": 5.0, "wind_dir_deg": WIND_TO_DEG, "rho_air": 1.225,
-        **overrides,
-    })
+    return BasicSail(
+        {
+            "airfoil_name": "NACA0012",
+            "area": 1.971,
+            "alpha_min": -179,
+            "alpha_max": 179,
+            "res": [1.8e5],
+            "wind_speed": 5.0,
+            "wind_dir_deg": WIND_TO_DEG,
+            "rho_air": 1.225,
+            **overrides,
+        }
+    )
 
 
 def tree(sail_rad: float, psi: float = 0.0) -> TFTree2D:
     """Transform tree with the boat on a heading and the sail trimmed."""
     tf = TFTree2D()
-    tf.add_frame(name="boat", parent="world",
-                 transform=Transform2D(x=0.0, y=0.0, c=math.cos(psi), s=math.sin(psi)))
-    tf.add_frame(name="sail", parent="boat",
-                 transform=Transform2D(x=0.0, y=0.0, c=math.cos(sail_rad), s=math.sin(sail_rad)))
+    tf.add_frame(name="boat", parent="world", transform=Transform2D(x=0.0, y=0.0, c=math.cos(psi), s=math.sin(psi)))
+    tf.add_frame(
+        name="sail", parent="boat", transform=Transform2D(x=0.0, y=0.0, c=math.cos(sail_rad), s=math.sin(sail_rad))
+    )
     return tf
 
 
@@ -51,8 +59,7 @@ def drive_coefficient(sail: BasicSail, twa_deg: float, u: float = 1.0) -> float:
     psi = beat(twa_deg)
     st = make_state(u=u, psi=psi)
     v_world = np.array([math.cos(psi), math.sin(psi)]) * u
-    aw = np.array([5.0 * math.cos(math.radians(WIND_TO_DEG)),
-                   5.0 * math.sin(math.radians(WIND_TO_DEG))]) - v_world
+    aw = np.array([5.0 * math.cos(math.radians(WIND_TO_DEG)), 5.0 * math.sin(math.radians(WIND_TO_DEG))]) - v_world
     q_area = 0.5 * 1.225 * float(np.dot(aw, aw)) * float(sail.p["area"])
     return max(sail.compute(st, tree(math.radians(-t), psi))[0] for t in range(5, 90, 5)) / q_area
 
@@ -184,7 +191,7 @@ class TestIntegratorContract:
             hub = SailboatHub("flingo_floty.yaml")
             psi = beat(45.0)
             st = State.from_array(np_.array([0.0, 0.0, math.cos(psi), math.sin(psi), 1.0, 0.0, 0.0]))
-            for _ in range(int(round(secs / dt))):
+            for _ in range(round(secs / dt)):
                 st = hub.step(st, dt, rk4_step, math.radians(20.0), 8.0)
             return np_.array([st.x, st.y, st.u, st.v, st.r])
